+++ src/test/suite/jsonc.test.ts
@@ -12,12 +12,63 @@ suite("JSON with comments", () => {
     );
   });
 
+  test("invalid char", () => {
+    expectHoverEqual(
+      new JsoncStringParser(),
+      new vscode.Position(0, 0),
+      "1\x19\n",
+      `"1\x19\\n"`
+    );
+  });
+
   test("string with escape", () => {
+    expectHoverEqual(
+      new JsoncStringParser(),
+      new vscode.Position(0, 0),
+      '1"23',
+      `"1\\"23"`
+    );
+
+    expectHoverEqual(
+      new JsoncStringParser(),
+      new vscode.Position(0, 0),
+      "1\\23",
+      `"1\\\\23"`
+    );
+
+    expectHoverEqual(
+      new JsoncStringParser(),
+      new vscode.Position(0, 0),
+      "1/23",
+      `"1\\/23"`
+    );
+
+    expectHoverEqual(
+      new JsoncStringParser(),
+      new vscode.Position(0, 0),
+      "1\b23",
+      `"1\\b23"`
+    );
+
+    expectHoverEqual(
+      new JsoncStringParser(),
+      new vscode.Position(0, 0),
+      "1\f23",
+      `"1\\f23"`
+    );
+
     expectHoverEqual(
       new JsoncStringParser(),
       new vscode.Position(0, 0),
       "1\n23",
       `"1\\n23"`
+    );
+
+    expectHoverEqual(
+      new JsoncStringParser(),
+      new vscode.Position(0, 0),
+      "1\r23",
+      `"1\\r23"`
     );
 
     expectHoverEqual(
@@ -30,9 +81,38 @@ suite("JSON with comments", () => {
     expectHoverEqual(
       new JsoncStringParser(),
       new vscode.Position(0, 0),
-      '1"23',
-      `"1\\"23"`
+      "1\u123423",
+      `"1\\u123423"`
     );
+  });
+
+  suite("bad escape", () => {
+    test("unicode", () => {
+      expectHoverEqual(
+        new JsoncStringParser(),
+        new vscode.Position(0, 0),
+        "1zzzz",
+        `"1\\uzzzz"`
+      );
+    });
+
+    test("unnecessary", () => {
+      expectHoverEqual(
+        new JsoncStringParser(),
+        new vscode.Position(0, 0),
+        "1a23",
+        `"1\\a23"`
+      );
+    });
+
+    test("unterminated", () => {
+      expectHoverEqual(
+        new JsoncStringParser(),
+        new vscode.Position(0, 0),
+        "1\\",
+        `"1\\`
+      );
+    });
   });
 
   test("string after comments", () => {
@@ -65,18 +145,48 @@ suite("JSON with comments", () => {
     );
   });
 
-  test("unclosed string", () => {
-    expectHoverEqual(
-      new JsoncStringParser(),
-      new vscode.Position(0, 0),
-      '1"',
-      `"1\\"`
-    );
-    expectHoverEqual(
-      new JsoncStringParser(),
-      new vscode.Position(0, 0),
-      '1"',
-      `"1\\"\n`
-    );
+  suite("unclosed string", () => {
+    test("end of input", () => {
+      expectHoverEqual(
+        new JsoncStringParser(),
+        new vscode.Position(0, 0),
+        "1\n",
+        `"1\\n`
+      );
+    });
+
+    test("end of line", () => {
+      expectHoverEqual(
+        new JsoncStringParser(),
+        new vscode.Position(0, 0),
+        "1\n",
+        `"1\\n\n`
+      );
+    });
+
+    test("end of line with escape", () => {
+      expectHoverEqual(
+        new JsoncStringParser(),
+        new vscode.Position(0, 0),
+        "1\\",
+        `"1\\`
+      );
+    });
+
+    test("escaped quote", () => {
+      expectHoverEqual(
+        new JsoncStringParser(),
+        new vscode.Position(0, 0),
+        '1"',
+        `"1\\"`
+      );
+
+      expectHoverEqual(
+        new JsoncStringParser(),
+        new vscode.Position(0, 0),
+        '1"',
+        `"1\\"\n`
+      );
+    });
   });
 });